+++ src/ros/path_plan/CMakeLists.txt
@@ -8,10 +8,11 @@ endif()
 # find dependencies
 find_package(ament_cmake REQUIRED)
 find_package(rclcpp REQUIRED)
-find_package(custom_types REQUIRED)
 find_package(nav_msgs REQUIRED)
-# uncomment the following section in order to fill in
-# further dependencies manually.
+find_package(geometry_msgs REQUIRED)
+
+find_package(Eigen3 REQUIRED)
+
 find_package(OpenCV)
 if(OpenCV_FOUND)
   MESSAGE(STATUS "OpenCV installation found")
@@ -21,8 +22,8 @@ else()
   MESSAGE(STATUS "OpenCV installation not found")
 endif()
 
-add_executable(main src/main.cpp src/path_plan.cpp src/WeightMap.cpp)
-ament_target_dependencies(main rclcpp custom_types nav_msgs)
+add_executable(main src/main.cpp src/path_plan/pathplan.cpp src/path_plan/WeightMap.cpp)
+ament_target_dependencies(main rclcpp nav_msgs geometry_msgs)
 
 if(OpenCV_FOUND)
   target_link_libraries(main ${OpenCV_LIBS})
@@ -30,7 +31,8 @@ endif()
 
 target_include_directories(main PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
+  $<INSTALL_INTERFACE:include>
+  PRIVATE ${EIGEN3_INCLUDE_DIR})
 target_compile_features(main PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
 
 install(TARGETS main
